requirements.txt
flask
python-dotenv
flask_mysqldb
flask-bcrypt
sqlalchemy
mysql-connector-python
mysql-connector-python-rf